rename load-settings to `environment`
