+++ packages/plugin-nillion/src/actions/common.ts
@@ -1,5 +1,5 @@
 import { elizaLogger, IAgentRuntime, Memory } from "@elizaos/core";
-import { loadConfig } from "./load-settings.ts";
+import { loadConfig } from "./environment";
 
 export async function validateNilDbConfig(
     runtime: IAgentRuntime,
